test: move to integration
