+++ tests/unit/docker/test_hub_build_level.py
@@ -33,39 +33,3 @@ def docker_image():
     images = client.images.list()
     image_name = list(filter(lambda image: _filter_repo_tag(image), images))[0]
     return image_name
-
-
-def test_hub_build_level_pass(monkeypatch, test_workspace, docker_image):
-    args = set_hub_build_parser().parse_args(
-        ['path/hub-mwu', '--push', '--host-info', '--test-level', 'EXECUTOR']
-    )
-    expected_failed_levels = []
-
-    _, failed_levels = HubIO(args)._test_build(
-        docker_image,
-        BuildTestLevel.EXECUTOR,
-        os.path.join(cur_dir, 'yaml/test-joint.yml'),
-        60000,
-        True,
-        JinaLogger('unittest'),
-    )
-
-    assert expected_failed_levels == failed_levels
-
-
-def test_hub_build_level_fail(monkeypatch, test_workspace, docker_image):
-    args = set_hub_build_parser().parse_args(
-        ['path/hub-mwu', '--push', '--host-info', '--test-level', 'FLOW']
-    )
-    expected_failed_levels = [BuildTestLevel.POD_DOCKER, BuildTestLevel.FLOW]
-
-    _, failed_levels = HubIO(args)._test_build(
-        docker_image,
-        BuildTestLevel.FLOW,
-        os.path.join(cur_dir, 'yaml/test-joint.yml'),
-        60000,
-        True,
-        JinaLogger('unittest'),
-    )
-
-    assert expected_failed_levels == failed_levels
